coverity: fix defects
- Explicit null dereferenced
- Overflowed constant
- Overflowed return value
- Resource leak

Defects: 392741, 392742, 392743, 392746, 392747, 392748, 392749, 392752,
         392754, 392755, 392756, 392758, 392759, 432787, 432788, 432790

+++ envelope/window.c
@@ -219,7 +219,7 @@ static int calc_address(struct AddressList *al, short cols, short *srows)
   struct ListHead slist = STAILQ_HEAD_INITIALIZER(slist);
   mutt_addrlist_write_list(al, &slist);
 
-  int rows = 1;
+  short rows = 1;
   int addr_len;
   int width_left = cols;
   struct ListNode *next = NULL;
